@@ -6,12 +6,11 @@ import android.app.DatePickerDialog;
 import android.app.Dialog;
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
+import androidx.fragment.app.DialogFragment;
 import android.widget.DatePicker;
 import android.widget.EditText;
 
 import io.adaptivecards.objectmodel.DateInput;
-import io.adaptivecards.objectmodel.DateTimePreparser;
 import io.adaptivecards.renderer.readonly.RendererUtil;
 
 import java.text.DateFormat;
