@@ -1,7 +1,7 @@
 package io.adaptivecards.adaptivecardssample.CustomObjects.CardElements;
 
 import android.content.Context;
-import android.support.v4.app.FragmentManager;
+import androidx.fragment.app.FragmentManager;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
