@@ -2,7 +2,7 @@ package io.adaptivecards.adaptivecardssample.CustomObjects.Media;
 
 import android.media.MediaDataSource;
 import android.os.Build;
-import android.support.annotation.RequiresApi;
+import androidx.annotation.RequiresApi;
 import android.util.Log;
 
 import java.io.ByteArrayOutputStream;
