@@ -4,16 +4,14 @@ package com.example.mobilechatapp;
 
 import android.app.Dialog;
 import android.content.Context;
-import android.support.v4.app.DialogFragment;
+import androidx.fragment.app.DialogFragment;
 
 import android.os.Bundle;
-import android.support.v4.app.FragmentManager;
+import androidx.fragment.app.FragmentManager;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.WindowManager;
-
-import com.example.mobilechatapp.R;
 
 import io.adaptivecards.objectmodel.HostConfig;
 import io.adaptivecards.objectmodel.ShowCardAction;
